feat: Add image field to user model
This commit adds an `image` field to the user model in the `user.js` file. The `image` field is of type `String` and allows users to store an image URL. This change enhances the functionality of the application by allowing users to have a profile picture associated with their account.