⚡️  feat: Add lazy loading for images and update webpack config
This commit adds lazy loading for images by configuring the `images` object in the `next.config.js` file. The `domains` property is updated to include 'lh3.googleusercontent.com', allowing images from this domain to be lazily loaded. Additionally, the webpack configuration is updated to enable top-level await, improving the performance of the application. The `next.config.mjs` file is also removed as it is no longer needed.